+++ packages/ui/src/mdx/components/pre.tsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { useContext, type FC, type ReactNode } from 'react';
-import { CodeSnippetsWrapper } from './code-snippets/wrapper';
+import { CodeSnippetsWrapper } from '../../code-snippets-wrapper/wrapper';
 import { FigureContext } from './figure-provider';
 
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions -- With an interface, we get this error in ./index: https://github.com/microsoft/TypeScript/issues/5711
 type PreProps = {
   children?: ReactNode;
   raw?: string;
-}
+};
 
 export const Pre: FC<PreProps> = ({ children, raw }) => {
   const context = useContext(FigureContext);
